Move drawer to garden related components folder
